fcm_service/pyproject.toml
[project]
name = "process-gpt-fcm-service"
version = "0.1.0"
description = "Process GPT FCM Service - Handles Firebase Cloud Messaging push notifications"
readme = "README.md"
requires-python = ">=3.12.3,<3.13"
dependencies = [
    "fastapi==0.115.12",
    "firebase-admin==6.9.0",
    "psutil==6.1.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "supabase==2.15.3",
    "uvicorn==0.34.3",
]